+++ packages/field/src/components/Options/index.tsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React, { useContext, useImperativeHandle } from 'react';
 import { Space, ConfigProvider } from 'antd';
 import type { ProFieldFC } from '../../index';
 
@@ -20,9 +20,11 @@ const addArrayKeys = (doms: React.ReactNode[]) =>
  *
  * @param
  */
-const FieldOptions: ProFieldFC = ({ text, mode: type, render, fieldProps }) => {
+const FieldOptions: ProFieldFC = ({ text, mode: type, render, fieldProps }, ref) => {
   const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
   const className = getPrefixCls('pro-field-option');
+
+  useImperativeHandle(ref, () => ({}));
 
   if (render) {
     const doms = render(text, { mode: type, ...fieldProps }, <></>) as unknown as React.ReactNode[];
@@ -52,4 +54,4 @@ const FieldOptions: ProFieldFC = ({ text, mode: type, render, fieldProps }) => {
   );
 };
 
-export default FieldOptions;
+export default React.forwardRef(FieldOptions);