+++ packages/layout/src/components/PageContainer/index.tsx
@@ -73,7 +73,7 @@ export type PageContainerProps = {
    *
    * @name 固钉的配置
    */
-  affixProps?: AffixProps;
+  affixProps?: Omit<AffixProps, 'children'>;
 
   /**
    * 只加载内容区域
@@ -267,6 +267,7 @@ const PageContainer: React.FC<PageContainerProps> = (props) => {
     affixProps,
     ghost = true,
     fixedHeader,
+    breadcrumbRender,
     ...restProps
   } = props;
   const value = useContext(RouteContext);
@@ -296,9 +297,15 @@ const PageContainer: React.FC<PageContainerProps> = (props) => {
       </>
     ) : null;
   }, [children, prefixedClassName, value.hasFooterToolbar]);
+
+  const memoBreadcrumbRender = useMemo(() => {
+    if (breadcrumbRender == false) return false;
+    return breadcrumbRender || restProps?.header?.breadcrumbRender;
+  }, [breadcrumbRender, restProps?.header?.breadcrumbRender]);
   const pageHeaderDom = (
     <ProPageHeader
       {...restProps}
+      breadcrumbRender={memoBreadcrumbRender}
       ghost={ghost}
       prefixCls={undefined}
       prefixedClassName={prefixedClassName}