+++ packages/layout/src/demos/api.tsx
@@ -137,6 +137,7 @@ export default () => {
                   blankTarget: true,
                 },
               ]}
+              copyright="2022 蚂蚁金服体验技术部出品"
             />
           ) : null
         }