+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.30](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.29...@ant-design/pro-descriptions@1.10.30) (2022-01-11)
+
+### Bug Fixes
+
+- antd themes less file path ([#4436](https://github.com/ant-design/pro-components/issues/4436)) ([a6b66b4](https://github.com/ant-design/pro-components/commit/a6b66b4968e78a9c8175d17d2fe5a922cf70630f))
+
+## [1.10.29](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.28...@ant-design/pro-descriptions@1.10.29) (2022-01-07)
+
+### Bug Fixes
+
+- **descriptions:** fix descriptions if is null ([28df05e](https://github.com/ant-design/pro-components/commit/28df05e12d3e534f1c118547a1b0a793c40d85c1))
+
+## [1.10.28](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.27...@ant-design/pro-descriptions@1.10.28) (2022-01-05)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.27](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.26...@ant-design/pro-descriptions@1.10.27) (2021-12-28)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.26](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.25...@ant-design/pro-descriptions@1.10.26) (2021-12-28)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [1.10.25](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.24...@ant-design/pro-descriptions@1.10.25) (2021-12-24)
 
 **Note:** Version bump only for package @ant-design/pro-descriptions