+++ packages/list/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.15](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.14...@ant-design/pro-list@1.21.15) (2022-01-11)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.14](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.13...@ant-design/pro-list@1.21.14) (2022-01-07)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.13](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.12...@ant-design/pro-list@1.21.13) (2022-01-05)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.12](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.11...@ant-design/pro-list@1.21.12) (2021-12-28)
+
+### Bug Fixes
+
+- **form:** fix ts error ([fbe853d](https://github.com/ant-design/pro-components/commit/fbe853d66aae90040ae0bc807c255fbf984611f2))
+
+## [1.21.11](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.10...@ant-design/pro-list@1.21.11) (2021-12-28)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [1.21.10](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.9...@ant-design/pro-list@1.21.10) (2021-12-24)
 
 **Note:** Version bump only for package @ant-design/pro-list