+++ packages/table/CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.62.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.62.1...@ant-design/pro-table@2.62.2) (2022-01-11)
+
+### Bug Fixes
+
+- **table:** fix defaultSelectedRowKeys no work error ([#4432](https://github.com/ant-design/pro-components/issues/4432)) ([4ee5dce](https://github.com/ant-design/pro-components/commit/4ee5dce23550b1175c988995afec8611cb78562a))
+
+## [2.62.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.62.0...@ant-design/pro-table@2.62.1) (2022-01-07)
+
+**Note:** Version bump only for package @ant-design/pro-table
+
+# [2.62.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.61.9...@ant-design/pro-table@2.62.0) (2022-01-05)
+
+### Bug Fixes
+
+- **table:** fix the param resetPageIndex is invalid in ProTable's method reload. ([#4397](https://github.com/ant-design/pro-components/issues/4397)) ([1169309](https://github.com/ant-design/pro-components/commit/11693090f876de639403bd6a1dfd4e57a4f81d8b))
+
+### Features
+
+- **ProTable:** support disabled setting's operation ([#4394](https://github.com/ant-design/pro-components/issues/4394)) ([89ef39f](https://github.com/ant-design/pro-components/commit/89ef39f1b2da7e43e6b221d289c0f8559d71e94f))
+
+## [2.61.9](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.61.8...@ant-design/pro-table@2.61.9) (2021-12-28)
+
+### Bug Fixes
+
+- **table:** fix table no rerender error ([5af32fb](https://github.com/ant-design/pro-components/commit/5af32fbf4bf68d93a7851cb3d10da339e6e26d19))
+
+## [2.61.8](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.61.7...@ant-design/pro-table@2.61.8) (2021-12-28)
+
+### Bug Fixes
+
+- **table:** Table 在 column 带有 request 情况下，搜索栏的表单由于没有设置 proFieldKey 导致在页面初始化会发出两个重 复的 request ([#4336](https://github.com/ant-design/pro-components/issues/4336)) ([219248b](https://github.com/ant-design/pro-components/commit/219248bbe80f3a3711524d668eac56626d131e89))
+
 ## [2.61.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.61.6...@ant-design/pro-table@2.61.7) (2021-12-24)
 
 **Note:** Version bump only for package @ant-design/pro-table